+++ docs/source/changes.rst
@@ -5,6 +5,8 @@ Changes
 ------
 
 * added :mod:`mgkit.graph` module
+* added function :func:`mgkit.io.gff.parse_glimmer3` and :meth:`mgkit.BaseGFFDict.from_glimmer3` to parse a GLIMMER3 output file and convert it to GFF annotation
+* added more tests
 
 0.1.11
 ------